tout marche

+++ ft_putptr_hexamin.c
@@ -6,7 +6,7 @@
 /*   By: itovar-n <marvin@42lausanne.ch>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/10 17:29:35 by itovar-n          #+#    #+#             */
-/*   Updated: 2022/12/20 17:44:42 by itovar-n         ###   ########.fr       */
+/*   Updated: 2022/12/21 15:28:04 by itovar-n         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 int	ft_putptr_hexamin(va_list ap)
 {
-	unsigned long	n;
-	int				result;
-	char			*base;
-	unsigned long	size;
+	unsigned long long	n;
+	int					result;
+	char				*base;
+	unsigned long long	size;
 
 	n = va_arg(ap, unsigned long long);
 	ft_putstr_fd("0x", 1);
@@ -29,7 +29,7 @@ int	ft_putptr_hexamin(va_list ap)
 		ft_putstr_fd("0", 1);
 		return (result);
 	}
-	size = ft_strlen(base);
+	size = (unsigned long long) ft_strlen(base);
 	ft_putnbr_base_fd(n, base, 1);
 	while (n >= size)
 	{
